+++ broker/settings.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from dynaconf import Dynaconf, Validator
 from dynaconf.validator import ValidationError
 from logzero import logger
+from broker.exceptions import ConfigurationError
 
 settings_file = "broker_settings.yaml"
 BROKER_DIRECTORY = Path()
@@ -14,6 +15,7 @@ if "BROKER_DIRECTORY" in os.environ:
         BROKER_DIRECTORY = envar_location
 
 settings_path = BROKER_DIRECTORY.joinpath("broker_settings.yaml")
+inventory_path = BROKER_DIRECTORY.joinpath("inventory.yaml")
 
 validators = [
     Validator("HOST_USERNAME", default="root"),
@@ -28,5 +30,6 @@ settings = Dynaconf(
 try:
     settings.validators.validate()
 except ValidationError as err:
-    logger.error(f"Configuration error in {settings_path.absolute()}: {err}")
-    sys.exit()
+    raise ConfigurationError(
+        f"Configuration error in {settings_path.absolute()}: {err.args[0]}"
+    )